Add files via upload
Adding some front end files, this includes proof of concept for sales page. The sales page will show a basic ticket image upon the clicking of 'Create Ticket'. This is just a proof of concept, so not super dynamic, as it is the same image every time you click the button.